@@ -1,10 +1,11 @@
 {
-  "name": "react-derivable",
+  "name": "@deskpro/react-derivable",
   "version": "0.2.0",
   "description": "Make React component react on changes in referened derivable values",
   "main": "lib/index.js",
   "scripts": {
-    "test": "make lint test"
+    "test": "make lint test",
+    "compile": "rm -rf lib/* && babel --loose es6.modules -d lib/ src/"
   },
   "repository": {
     "type": "git",
@@ -22,7 +23,7 @@
   "homepage": "https://github.com/andreypopp/react-derivable#readme",
   "peerDependencies": {
     "derivable": "1.0.0-beta11",
-    "react": ">= 0.14.0 < 16.0.0"
+    "react": ">= 0.14.0"
   },
   "devDependencies": {
     "babel-cli": "^6.10.1",
@@ -41,12 +42,13 @@
     "mocha": "^2.5.3",
     "pnpm": "^0.29.0",
     "power-assert": "^1.4.1",
-    "react": "^15.2.1",
-    "react-addons-test-utils": "^15.4.1",
-    "react-dom": "^15.2.1",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "sinon": "^1.17.4"
   },
   "dependencies": {
-    "invariant": "^2.2.1"
+    "invariant": "^2.2.1",
+    "prop-types": "^15.6.0"
   }
 }
