@@ -4,6 +4,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
 import {atom} from 'derivable';
 import * as Immutable from 'immutable';
 
@@ -531,38 +532,38 @@ function makeComponentDecoratorTestSuite(decorate) {
 
   it('preserves propTypes for class components', function() {
     class X extends React.Component {
-      static propTypes = {x: React.PropTypes.string};
+      static propTypes = {x: PropTypes.string};
       render() {
         return null;
       }
     }
-    assert(decorate(X).propTypes.x === React.PropTypes.string);
+    assert(decorate(X).propTypes.x === PropTypes.string);
   });
 
   it('preserves propTypes for functional components', function() {
     function X() {
       return null;
     }
-    X.propTypes = {x: React.PropTypes.string};
-    assert(decorate(X).propTypes.x === React.PropTypes.string);
+    X.propTypes = {x: PropTypes.string};
+    assert(decorate(X).propTypes.x === PropTypes.string);
   });
 
   it('preserves contextTypes for class components', function() {
     class X extends React.Component {
-      static contextTypes = {x: React.PropTypes.string};
+      static contextTypes = {x: PropTypes.string};
       render() {
         return null;
       }
     }
-    assert(decorate(X).contextTypes.x === React.PropTypes.string);
+    assert(decorate(X).contextTypes.x === PropTypes.string);
   });
 
   it('preserves contextTypes for functional components', function() {
     function X() {
       return null;
     }
-    X.contextTypes = {x: React.PropTypes.string};
-    assert(decorate(X).contextTypes.x === React.PropTypes.string);
+    X.contextTypes = {x: PropTypes.string};
+    assert(decorate(X).contextTypes.x === PropTypes.string);
   });
 
   it('preserves defaultProps for class components', function() {
